feat: split `useTheme` modules
