+++ src/ui/useTheme/isThemeDark.ts
@@ -1,0 +1,7 @@
+import { ThemeProps } from './ThemeProvider';
+
+export default function isThemeDark(
+  mode: ThemeProps['device']['mode']
+): mode is 'dark' {
+  return mode === 'dark';
+}
